working flow,remove runlike

+++ dffml/operation/deploy/deploy/operations.py
@@ -1,5 +1,7 @@
 import os
+import re
 import json
+import shlex
 import asyncio
 import pathlib
 from typing import List,Dict,Any
@@ -17,7 +19,6 @@ from dffml_feature_git.util.proc import check_output, create, stop
     outputs={"url":clone_git_repo.op.inputs["URL"]}
 )
 def get_url_from_payload(payload:Dict[str,Any]):
-    print(f" \n\n get_url_from_payload {payload} ")
     return {
         "url": payload["repository"]["clone_url"]
     }
@@ -26,7 +27,6 @@ def get_url_from_payload(payload:Dict[str,Any]):
 def check_if_default_branch(payload):
     pushed_branch = payload["ref"].split("/")[-1]
     default_branch = payload["repository"]["default_branch"]
-    print(f" is_default_branch:{default_branch==pushed_branch} ")
     return {
         "is_default_branch":(default_branch==pushed_branch)
     }
@@ -36,7 +36,6 @@ def get_image_tag(payload):
     url = payload["repository"]["html_url"] # eg:-"https://github.com/username/Hello-World",
     tag = url.split("/")
     tag = f"{tag[-2]}/{tag[-1]}"
-    print(f"image_tag:{tag} ")
     return {"image_tag":tag}
 
 @op(
@@ -46,14 +45,12 @@ def get_image_tag(payload):
 async def get_running_containers(tag):
     containers = await check_output("docker","ps","--filter",f"ancestor={tag}", "--format","{{.ID}}" )
     containers =[container for container in containers.strip().split("\n") if container ]
-    print(f"running containers : {containers}\n")
     return {
         "running_containers":containers
             }
 
 @op(inputs={"containers":docker_running_containers},outputs={"status":got_running_containers})
 def get_status_running_containes(containers):
-    print(f"Got running containers : {containers} : returning true")
     return{
         "status":True
     }
@@ -63,42 +60,61 @@ def get_status_running_containes(containers):
         "repo":clone_git_repo.op.outputs["repo"],
         "image_tag":docker_image_tag
         },
+    outputs = {"docker_commands":docker_commands}
+)
+async def parse_docker_commands(repo,image_tag):
+    docker_file = next(pathlib.Path(repo.directory).rglob("Dockerfile"))
+    docker_build_cmd = ["docker","build","-t",image_tag,str(docker_file.parent)]
+    with open(docker_file,'r') as f:
+        s = f.read()
+
+    # parsing run command from Dockerfile
+    # parses lines starting with "# docker run" to line ending with "# "
+    x = re.findall("(?:#[ ]*docker run )(?:.|\n)*(?:#[ ]*\n)",s)
+    if not x:
+        # handles case were `FROM` starts immediatly after `usage` comments,
+        # without blank comment in between
+        x = re.findall("((?:#[ ]*docker run )(?:.|\n)*)FROM",s)
+    #TODO raise error when docker file does not have run command in usage
+    x = x[0].replace('#','').strip()
+    docker_run_cmd = shlex.split(x)
+
+    return{
+        "docker_commands":{
+            "build":docker_build_cmd,
+            "run" :docker_run_cmd
+        }
+    }
+
+@op(
+    inputs={"docker_commands":docker_commands},
     outputs={"build_status":is_image_built},
     conditions=[is_default_branch,got_running_containers]
 )
-async def docker_build_image(repo,image_tag):
-    print(f"running docker build image")
-    # finding top most `Dockerfile`
-    print(f"\n\n repo : {repo}\n\n")
-    docker_file = next(pathlib.Path(repo.directory).rglob("Dockerfile"))
-    docker_file = str(docker_file.parents[0])
+async def docker_build_image(docker_commands):
     #building image
-    cmd_out = await check_output("docker","build","-t",image_tag,docker_file)
+    cmd_out = await check_output(*docker_commands["build"])
     build_status = "Successfully built" in cmd_out
-    print(f"cmd _out : {cmd_out}")
     return {"build_status":build_status}
 
 @op(
-    inputs={"tag":docker_image_tag,"containers":docker_running_containers},
+    inputs={"docker_commands":docker_commands,"containers":docker_running_containers},
     conditions = [is_image_built],
     outputs={}
     )
-async def restart_running_containers_by_tag(tag,containers):
-    # stop running containers and get commands used to start them
-    # and start it again
-    cmds=[]
+async def restart_running_containers(docker_commands,containers):
+    # if no containers are running ,start a fresh one else
+    # stop running containers and start it again with the new built
+    if not containers:
+        out = await check_output(*docker_commands["run"])
+        return
+
     for container in containers:
-        # get command used to start container
-        cmd = (await check_output("runlike",container)).strip()
-        print(f"Running : {cmd}")
         out = await check_output("docker","rm","-f",container)
         if not (out.strip()==container):
             print(f"out = {out}")
             print(f"container = {container}")
             raise Exception(f"Error when stopping container {container}")
-        # restart containers with newly build image
         # TODO : enforce starting in detach?
-        out = await check_output(*cmd.split(" "))
-        print(f"Out : {out}")
-
+        out = await check_output(*docker_commands["run"])
